// client/src/pageInfo/styled/HealthyDays.tsx
import styled from "styled-components";

export const SSummary = styled.div`
  font-size: 34px;
  color: #ffffff40;
`

export const SNumber = styled.div`
  color: #849A22;
  font-size: 72px;
  padding-top: 20px;
`;

export const SWrap = styled.div`
  display: flex;
  flex-direction: column;
  align-items: center;
`